refactor(FE): 프로젝트 탐색, 상세 페이지 컴포넌트명, path명 변경 (#152)
* refactor: 프로젝트 리스트 페이지 폴더 이름 변경

* refactor: 프로젝트 페이지 컴포넌트 이름 변경

* refactor: 프로젝트 탐색 페이지 path 변경

* refactor: 프로젝트 상세 페이지 path 변경

* refactor: 프로젝트 상세 페이지 폴더 이름 변경

* refactor: 프로젝트 상세 컴포넌트 이름 변경

* refactor: 프로젝트 카테고리 오타 수정
