+++ frontend/src/pages/project-detail/components/OverviewSection/OverviewSection.styled.ts
