+++ frontend/src/pages/project-detail/components/Post/Post.styled.ts
