+++ frontend/src/pages/project-detail/components/SectionTitle.styled.ts
