+++ frontend/src/pages/project-detail/components/TitleSection/TitleSection.styled.ts
