+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/CategoryFilterBox/CategoryFilterBox.styled.ts
