+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/FilterTrigger.styled.ts
