+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/TechStackFilterBox/TechStackSearchBar/TechStackSearchResult/TechStackSearchResult.styled.ts
