+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/TechStackFilterBox/TechStackSearchBar/TechStackSearchResult/TechStackSearchResult.tsx
@@ -1,6 +1,6 @@
 import type { TechStackKey } from "@domains/filter/techStack";
 import { useOutsideClick } from "@shared/hooks/useOutsideClick";
-import { useFilterParams } from "@/pages/list/hooks/useFilterParams";
+import { useFilterParams } from "@/pages/project-list/hooks/useFilterParams";
 import * as S from "./TechStackSearchResult.styled";
 
 interface TechStackSearchResultProps {
