+++ frontend/src/pages/project-list/ProjectSearchBar/ProjectSearchBar.tsx
